Se hizo la lista de turnos

+++ src/app/components/dashboard/business/employees/employees.component.ts
@@ -1,11 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import { AngularFireObject } from '@angular/fire/database';
+import { AngularFireDatabase, AngularFireObject } from '@angular/fire/database';
 import { Subscription } from 'rxjs';
 import { IEmployee } from 'src/app/core/models/employee.interface';
+import { ITurn } from 'src/app/core/models/turn.interface';
 import { AuthService } from 'src/app/services/auth.service';
 import { RepositoryService } from 'src/app/services/repository.service';
 import { UtilityService } from 'src/app/services/utility.service';
 import { AddEmployeeComponent } from './add-employee/add-employee.component';
+import { BEmployeeDetailsComponent } from './b-employee-details/b-employee-details.component';
 
 @Component({
   selector: 'app-employees',
@@ -20,10 +22,10 @@ export class EmployeesComponent implements OnInit {
   employeesSubscription: Subscription;
 
   uid: string;
-  keys: string[] = [];
 
   constructor(private authService: AuthService, 
     private repositoryService: RepositoryService<IEmployee[]>,
+    private angularFireDatabase: AngularFireDatabase,
     private utilityService: UtilityService) { }
 
   ngOnInit() {
@@ -39,10 +41,9 @@ export class EmployeesComponent implements OnInit {
       const data = result.payload.val();
 
       this.employeesList = [];
-      this.keys = [];
       for(let key in data){
+        data[key].key = key;
         this.employeesList.push(data[key]);
-        this.keys.push(key);
       }
     });
   }
@@ -51,20 +52,38 @@ export class EmployeesComponent implements OnInit {
     await this.utilityService.openModal(AddEmployeeComponent);
   }
 
-  async openDeleteConfirm(index: number){
+  async openDeleteConfirm(employee: IEmployee){
     await this.utilityService.presentAlertWithActions('Confirmar', 
     '¿Estás seguro de querer eliminar este empleado?',
-    ()=> this.deleteEmployee(index),
+    ()=> this.deleteEmployee(employee.key),
     ()=> this.cancelDeleteEmployee());
   }
 
-  async deleteEmployee(index: number){
-    const employeeKey = this.keys[index];
-    await this.repositoryService.deleteElement(`businessList/${this.uid}/employees/${employeeKey}`).then(async ()=>{
-      await this.utilityService.presentToast('Empleado eliminado correctamente', 'success-toast');
-    }).catch(async err=>{
-      await this.utilityService.presentToast('Ha ocurrido un error interno', 'error-toast');
+  async deleteEmployee(employeeKey: string){
+    const turnsRef: AngularFireObject<ITurn> = this.angularFireDatabase.object(`clientsInTurn/${this.uid}`);
+    const turns$ = turnsRef.snapshotChanges().subscribe(result=>{
+      const data = result.payload.val();
+      for(let key in data){
+        if(data[key].employeeKey === employeeKey){
+          this.utilityService.presentToast('Este empletado tiene turnos, no puedes eliminarlo', 'error-toast');
+          turns$.unsubscribe();
+          break;
+        }else{
+          this.repositoryService.deleteElement(`businessList/${this.uid}/employees/${employeeKey}`).then(async ()=>{
+            await this.utilityService.presentToast('Empleado eliminado correctamente', 'success-toast');
+            turns$.unsubscribe();
+          }).catch(async err=>{
+            await this.utilityService.presentToast('Ha ocurrido un error interno', 'error-toast');
+            turns$.unsubscribe();
+          });
+          break;
+        }
+      }
     });
+  }
+
+  async goToEmployeeDetails(employee: IEmployee){
+    await this.utilityService.openModal(BEmployeeDetailsComponent, employee);
   }
 
   cancelDeleteEmployee(){
